requirements/development.txt
selenium
ddt
python-dotenv
pytest
pylint
sphinx
furo
